requirements.txt
git+https://github.com/huggingface/transformers.git 
deepspeed
accelerate
sentencepiece
langchain
torch
bitsandbytes
langchain_community
"unstructured[pdf]"
sentence-transformers
faiss-gpu
pypdf
langchain_community
langchain_chroma